Corregi la norminette, añadi ft_lstiter y el Makefile

+++ ft_split.c
@@ -6,7 +6,7 @@
 /*   By: lpin < lpin@student.42malaga.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/28 16:54:24 by lpin              #+#    #+#             */
-/*   Updated: 2023/10/01 17:55:08 by lpin             ###   ########.fr       */
+/*   Updated: 2023/10/10 13:15:36 by lpin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,7 +59,7 @@ char	**ft_split(char const *s, char c)
 	char	*to_split;
 	char	**splitted;
 	size_t	substr_qty;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	to_split = (char *)s;
